#232: Remove terasoluna-gfw-functionaltest-infra-mybatis2 (merge into domain)
