+++ terasoluna-gfw-functionaltest-domain/src/main/java/org/terasoluna/gfw/functionaltest/domain/service/queryescape/DatabaseMetaInfoServiceMybatis2Impl.java
@@ -13,7 +13,7 @@
  * either express or implied. See the License for the specific language
  * governing permissions and limitations under the License.
  */
-package org.terasoluna.gfw.functionaltest.infra.repository.queryescape;
+package org.terasoluna.gfw.functionaltest.domain.service.queryescape;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
