--- terasoluna-gfw-functionaltest-infra-mybatis2/pom.xml
@@ -1,70 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-    <artifactId>terasoluna-gfw-functionaltest-infra-mybatis2</artifactId>
-    <packaging>jar</packaging>
-    <parent>
-        <groupId>org.terasoluna.gfw.functionaltest</groupId>
-        <artifactId>terasoluna-gfw-functionaltest-parent</artifactId>
-        <version>1.0.3-SNAPSHOT</version>
-        <relativePath>../terasoluna-gfw-functionaltest-parent/pom.xml</relativePath>
-    </parent>
-    <name>TERASOLUNA Global Framework - Functionality Test Application - MyBatis2 Layer</name>
-    <description>MyBatis2 dependent layer of Functionality Test Application using TERASOLUNA Global Framework</description>
-    <dependencies>
-        <dependency>
-            <groupId>${project.groupId}</groupId>
-            <artifactId>terasoluna-gfw-functionaltest-domain</artifactId>
-            <version>${project.version}</version>
-            <exclusions>
-                <exclusion>
-                    <groupId>org.terasoluna.gfw</groupId>
-                    <artifactId>terasoluna-gfw-jpa</artifactId>
-                </exclusion>
-                <exclusion>
-                    <groupId>org.terasoluna.gfw</groupId>
-                    <artifactId>terasoluna-gfw-security-core</artifactId>
-                </exclusion>
-            </exclusions>
-        </dependency>
-
-        <!-- == Begin TERASOLUNA == -->
-        <dependency>
-            <groupId>org.terasoluna.gfw</groupId>
-            <artifactId>terasoluna-gfw-mybatis2</artifactId>
-        </dependency>
-        <!-- == End TERASOLUNA == -->
-
-        <!-- == Begin Database == -->
-        <dependency>
-            <groupId>com.h2database</groupId>
-            <artifactId>h2</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.postgresql</groupId>
-            <artifactId>postgresql</artifactId>
-        </dependency>
-        <!-- <dependency> -->
-        <!-- <groupId>com.oracle</groupId> -->
-        <!-- <artifactId>ojdbc6</artifactId> -->
-        <!-- </dependency> -->
-        <!-- == End Database == -->
-
-        <!-- == Begin Unit Test == -->
-        <dependency>
-            <groupId>junit</groupId>
-            <artifactId>junit</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.mockito</groupId>
-            <artifactId>mockito-all</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.springframework</groupId>
-            <artifactId>spring-test</artifactId>
-        </dependency>
-        <!-- == End Unit Test == -->
-    </dependencies>
-
-</project>
