--- terasoluna-gfw-functionaltest-infra-mybatis2/src/main/resources/META-INF/mybatis/config/sqlMapConfig.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!DOCTYPE sqlMapConfig 
-            PUBLIC "-//ibatis.apache.org//DTD SQL Map Config 2.0//EN"
-            "http://ibatis.apache.org/dtd/sql-map-config-2.dtd">
-<sqlMapConfig>
-    <settings useStatementNamespaces="true" />
-</sqlMapConfig>
