--- terasoluna-gfw-functionaltest-infra-mybatis2/src/main/resources/META-INF/spring/terasoluna-gfw-functionaltest-infra-mybatis.xml
@@ -1,24 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<beans xmlns="http://www.springframework.org/schema/beans"
-    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xmlns:jpa="http://www.springframework.org/schema/data/jpa"
-    xmlns:util="http://www.springframework.org/schema/util"
-    xsi:schemaLocation="
-        http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd
-        http://www.springframework.org/schema/util http://www.springframework.org/schema/util/spring-util.xsd
-        ">
-
-	<bean id="sqlMapClient"
-        class="org.springframework.orm.ibatis.SqlMapClientFactoryBean">
-        <property name="configLocations"
-            value="classpath*:/META-INF/mybatis/config/*sqlMapConfig.xml" />
-        <property name="mappingLocations"
-            value="classpath*:/META-INF/mybatis/sql/**/*-sqlmap.xml" />
-        <property name="dataSource" ref="dataSource" />
-    </bean>
-    
-    <bean id="queryDAO" class="jp.terasoluna.fw.dao.ibatis.QueryDAOiBatisImpl">
-        <property name="sqlMapClient" ref="sqlMapClient" />
-    </bean>
-
-</beans>
