--- terasoluna-gfw-functionaltest-infra-mybatis2/src/test/resources/test-context.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<beans xmlns="http://www.springframework.org/schema/beans"
-    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd">
-
-    <import resource="classpath:META-INF/spring/terasoluna-gfw-functionaltest-infra-mybatis.xml" />
-
-    <bean class="org.springframework.jdbc.core.JdbcTemplate">
-        <property name="dataSource" ref="dataSource" />
-    </bean>
-</beans>
